Changed expected to int as only using ints for prices, started on checkout class

+++ CheckoutUnitTests/UnitTest1.cs
@@ -5,7 +5,7 @@ namespace CheckoutUnitTests
     public class UnitTest1
     {
         [Fact]
-        public void Test1(List<string> items,decimal expected)
+        public void Test1(List<string> items,int expected)
         {
             //Scan items
             ICheckout checkout = new Checkout();
